+++ src/main/java/duke/Duke.java
@@ -1,11 +1,7 @@
-import java.io.File;
+package duke;
+
 import java.io.FileNotFoundException;
-import java.io.FileWriter;
 import java.io.IOException;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.util.Scanner;
-import java.util.ArrayList;
 
 public class Duke {
     private Storage storage;
